Added iteration over subcollection Ticket

+++ entity/ticket.go
@@ -3,5 +3,5 @@ package entity
 //Ticket struct containing the details of a ticket
 type Ticket struct {
 	TicketType         string `json:"ticketType"`
-	NumberTicketsAvail int    `json:"numberTicketsAvail"`
+	NumberTicketsAvail int64  `json:"numberTicketsAvail"`
 }
